+++ Operations.h
@@ -4,6 +4,6 @@
 class Operations
 {
 public:
-	int compare(const Shape& obj_1, const Shape& obj_2);
-	int radiusDifference(const Shape& obj_1, const Shape& obj_2);
+	double compare(const Shape& obj_1, const Shape& obj_2);
+	double radiusDifference(const Shape& obj_1, const Shape& obj_2);
 };
